chore: rename collapse
