--- src/components/Accordion/index.js
@@ -1,2 +1,0 @@
-import Accordion from "./Accordion";
-export default Accordion;
