+++ src/components/Collapse/index.js
@@ -1,0 +1,2 @@
+import Accordion from "./Collapse";
+export default Accordion;
