+++ src/components/index.js
@@ -1,5 +1,5 @@
 export * from "./Glyphs";
-export { default as Accordion } from "./Accordion";
+export { default as Accordion } from "./Collapse";
 export { default as AppFooter } from "./AppFooter";
 export { default as Breadcrumbs } from "./Breadcrumbs";
 export { default as Button } from "./Button";
